Possible to place the A marker after moving skiArea
